rseq: Split out rseq.h from sched.h
mainline inclusion
from mainline-v6.8-rc1
category: performace

We're trying to get sched.h down to more or less just types only, not
code - rseq can live in its own header.

This helps us kill the dependency on preempt.h in sched.h.

Conflicts:
	fs/exec.c
	include/linux/sched.h
	kernel/fork.c
(cherry picked from commit 932562a6045ed613d45bd100db37114273c22077)
