+++ kernel/fork.c
@@ -102,6 +102,7 @@
 #include <linux/user_events.h>
 #include <linux/iommu.h>
 #include <linux/tick.h>
+#include <linux/rseq.h>
 
 #include <asm/pgalloc.h>
 #include <linux/uaccess.h>
